+++ src/main/java/shop/wannab/book_service/book/dto/BookInfoForOrderProjection.java
@@ -1,0 +1,8 @@
+package shop.wannab.book_service.book.dto;
+
+public interface BookInfoForOrderProjection {
+    long getBookId();
+    String getTitle();
+    int getOriginPrice();
+    int getSalesPrice();
+}
