+++ src/main/java/shop/wannab/book_service/book/dto/CartItem.java
@@ -1,0 +1,13 @@
+package shop.wannab.book_service.book.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+public class CartItem {
+    private long bookId;
+    private int quantity;
+}
