+++ src/main/java/shop/wannab/book_service/book/dto/OrderItemListDto.java
@@ -1,0 +1,14 @@
+package shop.wannab.book_service.book.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+public class OrderItemListDto {
+    private List<CartItem> orderItems;
+}
