+++ src/main/java/shop/wannab/book_service/book/repository/BookRedisRepository.java
@@ -1,0 +1,5 @@
+package shop.wannab.book_service.book.repository;
+
+public interface BookRedisRepository {
+    Integer getBookStock(long bookId);
+}
