+++ src/main/java/shop/wannab/book_service/category/entity/Category.java
@@ -1,0 +1,41 @@
+package shop.wannab.book_service.category.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import java.util.ArrayList;
+import java.util.List;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Entity
+@Getter
+@Setter
+@NoArgsConstructor
+public class Category {
+
+    @Id
+    private Long id;
+
+    @Column(nullable = false)
+    private String name;
+
+    @ManyToOne(fetch= FetchType.LAZY)
+    @JoinColumn(name= "parent_id")
+    private Category parent;
+
+    //자식 조회할때 사용
+    @OneToMany(mappedBy = "parent",fetch = FetchType.EAGER)
+    private List<Category> children = new ArrayList<>();
+
+    public Category(Long id, String name, Category parent) {
+        this.id = id;
+        this.name = name;
+        this.parent = parent;
+    }
+}
