+++ src/main/java/shop/wannab/book_service/category/repository/CategoryRepository.java
@@ -1,0 +1,9 @@
+package shop.wannab.book_service.category.repository;
+
+import java.util.List;
+import org.springframework.data.jpa.repository.JpaRepository;
+import shop.wannab.book_service.category.entity.Category;
+
+public interface CategoryRepository extends JpaRepository<Category, Long> {
+    List<Category> findByParentIsNull();
+}
