--- src/main/java/shop/wannab/book_service/constants/Constants.java
@@ -1,5 +1,0 @@
-package shop.wannab.book_service.constants;
-
-public abstract class Constants {
-    public static final String BOOK_STOCK_KEY = "book_stock";
-}
