--- src/main/java/shop/wannab/book_service/entity/Book.java
@@ -1,24 +1,0 @@
-package shop.wannab.book_service.entity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-public class Book {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long bookId;
-    private String title;
-    private int originPrice;
-    private int salesPrice;
-    private int stock;
-    private boolean isOnSale;
-}
