--- src/main/java/shop/wannab/book_service/entity/dto/BookInfoForOrderProjection.java
@@ -1,8 +1,0 @@
-package shop.wannab.book_service.entity.dto;
-
-public interface BookInfoForOrderProjection {
-    long getBookId();
-    String getTitle();
-    int getOriginPrice();
-    int getSalesPrice();
-}
