--- src/main/java/shop/wannab/book_service/entity/dto/CartItem.java
@@ -1,13 +1,0 @@
-package shop.wannab.book_service.entity.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-public class CartItem {
-    private long bookId;
-    private int quantity;
-}
