--- src/main/java/shop/wannab/book_service/entity/dto/OrderItemListDto.java
@@ -1,14 +1,0 @@
-package shop.wannab.book_service.entity.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-public class OrderItemListDto {
-    private List<CartItem> orderItems;
-}
