--- src/main/java/shop/wannab/book_service/entity/dto/OrderItemValidationError.java
@@ -1,11 +1,0 @@
-package shop.wannab.book_service.entity.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class OrderItemValidationError {
-    private long bookId;
-    private String message;
-}
