+++ src/main/java/shop/wannab/book_service/global/constants/Constants.java
@@ -1,0 +1,5 @@
+package shop.wannab.book_service.global.constants;
+
+public abstract class Constants {
+    public static final String BOOK_STOCK_KEY = "book_stock";
+}
