--- src/main/java/shop/wannab/book_service/repository/BookRedisRepository.java
@@ -1,5 +1,0 @@
-package shop.wannab.book_service.repository;
-
-public interface BookRedisRepository {
-    Integer getBookStock(long bookId);
-}
