--- src/main/java/shop/wannab/book_service/repository/CategoryRepository.java
@@ -1,9 +1,0 @@
-package shop.wannab.book_service.repository;
-
-import java.util.List;
-import org.springframework.data.jpa.repository.JpaRepository;
-import shop.wannab.book_service.entity.category.Category;
-
-public interface CategoryRepository extends JpaRepository<Category, Long> {
-    List<Category> findByParentIsNull();
-}
