--- src/main/java/shop/wannab/book_service/service/CategoryService.java
@@ -1,32 +1,0 @@
-package shop.wannab.book_service.service;
-
-import java.util.ArrayList;
-import java.util.List;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import shop.wannab.book_service.entity.category.Category;
-import shop.wannab.book_service.entity.category.dto.CategoryHierarchyDto;
-import shop.wannab.book_service.repository.CategoryRepository;
-
-
-@Service
-@RequiredArgsConstructor
-public class CategoryService {
-    private final CategoryRepository categoryRepository;
-
-    //재귀적으로 카테고리를 반환해주는 코드
-    @Transactional(readOnly = true)
-    public List<CategoryHierarchyDto> getCategoryHierarchy() {
-        //최상위 카테고리 호출
-        List<Category> rootCategories = categoryRepository.findByParentIsNull();
-
-        List<CategoryHierarchyDto> categoryHierarchyDtoList = new ArrayList<>();
-
-        for (Category rootCategory : rootCategories) {
-            CategoryHierarchyDto dto = new CategoryHierarchyDto(rootCategory);
-            categoryHierarchyDtoList.add(dto);
-        }
-        return categoryHierarchyDtoList;
-    }
-}
